requirements.txt
# Django and WebSocket dependencies
Django==4.2.7
channels==4.0.0
channels-redis==4.1.0
djangorestframework==3.14.0
django-cors-headers==4.3.1

# FastAPI and authentication dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.25.2
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data pipeline dependencies
redis==5.0.1
celery==5.3.4
kafka-python==2.0.2
pandas==2.1.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Development and visualization dependencies
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.21.1
matplotlib==3.7.2 